facts: do not cache merged var unless requested
The module is automatically adding merged variables to the facts cache - this
may have undesired consequences.  One of them being that the module issues the
warning about overwriting the target variable on any subsequent run.

Align the behavior with the set_fact module, where the `cacheable` argument
determines whether the fact gets stored in the facts cache - and defaults to False.

Original behavior (storing the merged variable in the facts cache) be be achieved by passing

   cacheable: True

E.g.:

   name: Merge user vars
   merge_vars:
     suffix_to_merge: users__to_merge
     merged_var_name: merged_users
     expected_type: 'dict'
     cacheable: True

But not caching should be the expected behavior most of the time.