idempotence: sort vars before merging
Python dicts are unordered.  And we get the list of variables to merge from a dict - so in arbitrary order.

That can have effect on the merged variables - and can produce spurious changes
on subsequent runs of the same playbook with the same input variables.
(E.g., when merging a dict of lists, the entries in the inner lists could be ordered
in different ways depending on the order in which the variables were merged)

This can get worked around by playbooks using this module by sorting the lists
before using them ... but that is sub-optimal.

Improve the situation by sorting the list of variable names before merging.

And, in recursive merging, also sort the keys of any dict being merged.